Update newYearChaos.swift
changed algorithm. Can't just sum the +ve diffs as that doesn't handle the cases when a person swaps with another person of +ve diffs. Instead need to compare each person to every person in front of them except: 1) cases where diff > 2 (impossible) and 2) cases where the person in the outer loop is the highest possible remaining person (therefore it is trivially true that no-one in front of that person could be a higher # and so is already at the back of the remaining people)